Synchronized with last Erwan's modification, 4 months ago

+++ src/main/java/org/cts/op/transformation/grids/BleggGeographicGrid.java
@@ -269,7 +269,7 @@ public class BleggGeographicGrid extends GeographicGrid {
                 int[] blockClone = new int[groupSize];
                 System.arraycopy(block, 0, blockClone, 0, groupSize);
                 Arrays.sort(blockClone);
-                int maxAbs = (int) Math.max(Math.abs(blockClone[0]),
+                int maxAbs = Math.max(Math.abs(blockClone[0]),
                         Math.abs(blockClone[blockClone.length - 1]));
                 // Nombre de bits necessaires = log2(maxAbs)
                 // + 1 pour compenser l'arrondi
@@ -389,7 +389,7 @@ public class BleggGeographicGrid extends GeographicGrid {
          */
         public void setValue(int value, int n, int size) {
             int powerOf2 = 1;
-            int abs = (int) Math.abs(value);
+            int abs = Math.abs(value);
             for (int i = 0; i < (size - 1); i++) {
                 if ((abs & powerOf2) == powerOf2) {
                     bitSet.set(n * size + i);
